# swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API documentation for the Exakta Laboratories turn taking application",
        "version": "1.0.0",
        "title": "API Turnero",
        "contact": {
            "email": "exakta@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "schemes": ["http"],
    "host": "localhost:4000",
    "basePath": "/api",
...
    "paths" : {
...
        "/roles": {
            "get": {
                "tags": ["Roles"],
                "summary" : "Get roles",
                "description": "Get roles",
                "produces": ["application/json"],
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "statusCode" : {
                                    "type": "number"
                                },
                                "body" : {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/definitions/RoleRes"
                                    }
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "example": {
                                "statusCode": 200,
                                "body": [
                                    {
                                        "_id": "618ea8befe20824c4f751010",
                                        "name": "Admin",
                                        "createdAt": "2021-11-12T17:47:42.286Z",
                                        "updatedAt": "2021-11-12T17:47:42.286Z",
                                        "__v": 0
                                    }
                                ],
                                "message": "Successful response"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "statusCode" : {
                                    "type": "number"
                                },
                                "body" : {
                                    "$ref": "#/definitions/InvalidRes"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "example": {
                                "statusCode": 401,
                                "body": {
                                    "message": "unauthorized|TokenExpired"
                                },
                                "message": "Client error"
                            }
                        }
                    },
                    "5XX": {
                        "description": "Server error",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "statusCode" : {
                                    "type": "number"
                                },
                                "body" : {
                                    "$ref": "#/definitions/InvalidRes"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "example": {
                                "statusCode": 500,
                                "body": {
                                    "message": "Message error."
                                },
                                "message": "Server error"
                            }
                        }
                    }
                },
                "security":[ { "api_key":[] } ]
            },
            "post": {
                "tags": ["Roles"],
                "summary" : "Create a role",
                "description": "Create a role",
                "produces": ["application/json"],
                "parameters": [
                    {
                        "in": "body",
                        "name": "role",
                        "description": "Role object",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/RoleReq"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "statusCode" : {
                                    "type": "number"
                                },
                                "body" : {
                                    "$ref": "#/definitions/RoleRes"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "example": {
                                "statusCode": 201,
                                "body": {
                                    "_id": "618ea8befe20824c4f751010",
                                    "name": "Admin",
                                    "createdAt": "2021-11-12T17:47:42.286Z",
                                    "updatedAt": "2021-11-12T17:47:42.286Z",
                                    "__v": 0
                                },
                                "message": "Successful response"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "statusCode" : {
                                    "type": "number"
                                },
                                "body" : {
                                    "$ref": "#/definitions/InvalidRes"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "example": {
                                "statusCode": 401,
                                "body": {
                                    "message": "unauthorized|TokenExpired"
                                },
                                "message": "Client error"
                            }
                        }
                    },
                    "5XX": {
                        "description": "Server error",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "statusCode" : {
                                    "type": "number"
                                },
                                "body" : {
                                    "$ref": "#/definitions/InvalidRes"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "example": {
                                "statusCode": 500,
                                "body": {
                                    "message": "Message error."
                                },
                                "message": "Server error"
                            }
                        }
                    }
                },
                "security":[ { "api_key":[] } ]
            }
        },
        "/roles/{role}": {
            "delete": {
                "tags": ["Roles"],
                "summary" : "Delete a role",
                "description": "Delete a role",
                "produces": ["application/json"],
                "parameters": [
                    {
                        "in": "path",
                        "name": "role",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "Vigia"
                        },
                        "description": "Name of the role"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "statusCode" : {
                                    "type": "number"
                                },
                                "body" : {
                                    "type": "object",
                                    "message": {
                                        "type": "string"
                                    }
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "example": {
                                "statusCode": 200,
                                "body": {
                                    "message": "Successfully deleted user."
                                },
                                "message": "Successful response"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "statusCode" : {
                                    "type": "number"
                                },
                                "body" : {
                                    "$ref": "#/definitions/InvalidRes"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "example": {
                                "statusCode": 401,
                                "body": {
                                    "message": "unauthorized|TokenExpired"
                                },
                                "message": "Client error"
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "statusCode" : {
                                    "type": "number"
                                },
                                "body" : {
                                    "type": "object"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "example": {
                                "statusCode": 404,
                                "body": {},
                                "message": "Client error"
                            }
                        }
                    },
                    "5XX": {
                        "description": "Server error",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "statusCode" : {
                                    "type": "number"
                                },
                                "body" : {
                                    "$ref": "#/definitions/InvalidRes"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "example": {
                                "statusCode": 500,
                                "body": {
                                    "message": "Message error."
                                },
                                "message": "Server error"
                            }
                        }
                    }
                },
                "security":[ { "api_key":[] } ]
            }
        },
...
    }, 
    "definitions": {
...
        "InvalidRes": {
            "type": "object",
            "properties": {
                "message":  {
                    "type": "string"
                }
            }
        },
...
        "RoleRes": {
            "type": "object",
            "properties": {
                "_id":  {
                    "type": "string"
                },
                "name":  {
                    "type": "string"
                },
                "createdAt":  {
                    "type": "string"
                },
                "updatedAt":  {
                    "type": "string"
                },
                "__v":  {
                    "type": "number"
                }
            }
        },
        "RoleReq": {
            "type": "object",
            "properties": {
                "name":  {
                    "type": "string"
                }
            },
            "required": [ 
                "name"
            ],
            "example": {
                "name":"Vigia"
            }
        },
...
    },
    "securityDefinitions": {
        "api_key": {
            "type": "apiKey",
            "name": "auth",
            "in": "header"
        }
    }
}